+++ DesignerRHN/DesignerRHN-ejb/src/java/Entidades/TiposEntidades.java
@@ -6,6 +6,7 @@ package Entidades;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Collection;
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -40,7 +41,7 @@ public class TiposEntidades implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "SECUENCIA")
-    private BigDecimal secuencia;
+    private BigInteger secuencia;
     @Basic(optional = false)
     @NotNull
     @Column(name = "CODIGO")
@@ -60,21 +61,21 @@ public class TiposEntidades implements Serializable {
     public TiposEntidades() {
     }
 
-    public TiposEntidades(BigDecimal secuencia) {
+    public TiposEntidades(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
 
-    public TiposEntidades(BigDecimal secuencia, short codigo, String nombre) {
+    public TiposEntidades(BigInteger secuencia, short codigo, String nombre) {
         this.secuencia = secuencia;
         this.codigo = codigo;
         this.nombre = nombre;
     }
 
-    public BigDecimal getSecuencia() {
+    public BigInteger getSecuencia() {
         return secuencia;
     }
 
-    public void setSecuencia(BigDecimal secuencia) {
+    public void setSecuencia(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
